Casi terminada parte 3

+++ src/main/java/com/university/aprovedCheckers/CheckApproved.java
@@ -7,5 +7,5 @@ import com.university.objetos.evaluations.Evaluations;
 import java.util.List;
 
 public interface CheckApproved {
-    boolean checkApproved(List<Evaluations> exams, int nota);
+    boolean checkApproved(List<Evaluations> exams, float nota);
 }
